update runsite command and default project config
I’m not sure if I should even use the runsite command or the project
config file. In production I have a system service command that has all
of the config information templated into it. `runsite` is probably only
good as a clone of that I can quickly test if I’m facing server issues.

For now, I’ve updated it to work with my Koa site template.